# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Resource-Service is service for storing kubernetes resources in Mongo DB.",
    "title": "Resource-Service",
    "version": "0.0.1"
  },
  "paths": {
...
    "/namespaces/{namespace}/configmaps": {
      "get": {
        "tags": [
          "ConfigMap"
        ],
        "summary": "Get configmaps list.",
        "operationId": "GetConfigMapsListHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "configmaps list",
            "schema": {
              "$ref": "#/definitions/ConfigMapsResponse"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "ConfigMap"
        ],
        "summary": "Create configmap.",
        "operationId": "CreateConfigMapHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ConfigMap"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "configmap created",
            "schema": {
              "$ref": "#/definitions/ResourceConfigMap"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "delete": {
        "tags": [
          "ConfigMap"
        ],
        "summary": "Delete configmap.",
        "operationId": "DeleteAllConfigMapsHandler",
        "parameters": [
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "all configmaps in namespace deleted"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "private"
      }
    },
    "/namespaces/{namespace}/configmaps/{configmap}": {
      "get": {
        "tags": [
          "ConfigMap"
        ],
        "summary": "Get configmaps list.",
        "operationId": "GetConfigMapHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "configmap",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "configmap",
            "schema": {
              "$ref": "#/definitions/ResourceConfigMap"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "delete": {
        "tags": [
          "ConfigMap"
        ],
        "summary": "Delete configmap.",
        "operationId": "DeleteConfigMapHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "configmap",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "configmap deleted"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
...
  },
  "definitions": {
...
    "ConfigMap": {
      "description": "ConfigMap -- model for config map",
      "type": "object",
      "required": [
        "name",
        "data"
      ],
      "properties": {
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "data": {
          "$ref": "#/definitions/ConfigMapData"
        },
        "deleted_at": {
          "description": "delete date in RFC3339 format",
          "type": "string",
          "x-go-name": "DeletedAt"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "ConfigMapData": {
      "description": "ConfigMapData -- model for config map data",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "git.containerum.net/ch/resource-service/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "ConfigMapsResponse": {
      "description": "ConfigMapsResponse -- configmap response",
      "type": "object",
      "properties": {
        "config_maps": {
          "$ref": "#/definitions/ListCM"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/pkg/models/configmap"
    },
...
    "ListCM": {
      "description": "ListCM -- ConfigMaps list",
      "type": "array",
      "items": {
        "$ref": "#/definitions/ResourceConfigMap"
      },
      "x-go-package": "git.containerum.net/ch/resource-service/pkg/models/configmap"
    },
...
    "ResourceConfigMap": {
      "description": "ResourceConfigMap --  model for ConfigMap for resource-service db",
      "type": "object",
      "required": [
        "name",
        "data"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "data": {
          "$ref": "#/definitions/ConfigMapData"
        },
        "deleted": {
          "type": "boolean",
          "x-go-name": "Deleted"
        },
        "deleted_at": {
          "description": "delete date in RFC3339 format",
          "type": "string",
          "x-go-name": "DeletedAt"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        },
        "namespaceid": {
          "type": "string",
          "x-go-name": "NamespaceID"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/pkg/models/configmap"
    },
...
    "errOAIGen": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "description": "Fields -- represents additional key-value fields of error",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-package": "github.com/containerum/cherry"
        },
        "id": {
          "$ref": "#/definitions/errOaiGenId"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-package": "github.com/containerum/cherry"
    },
    "errOaiGenId": {
      "description": "ErrID -- represents unique error ID",
      "type": "object",
      "properties": {
        "kind": {
          "description": "ErrKind -- represents kind of error",
          "type": "integer",
          "format": "uint64",
          "x-go-package": "github.com/containerum/cherry"
        },
        "sid": {
          "description": "ErrSID -- represents service ID of error",
          "type": "string",
          "x-go-package": "github.com/containerum/cherry"
        }
      },
      "x-go-gen-location": "models",
      "x-go-package": "github.com/containerum/cherry"
    }
  },
...
}